Added 'move' command
arguments:
- ( number )
- ( number , number )
- swap
- last

( number ): Move the selected song to the top of the queue.
Example: move 5

( number, number ): Move the selected song to the provided position.
Example: move 5 2

swap: Swap track positions in the queue.
Example: move swap 8 2

last: Move the last track in the queue to the top.
Example: move last